pyproject.toml
[build-system]
# we use the hatch build system
# https://hatch.pypa.io/latest/config/build/
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-3rdparty-mediawiki"
description="Wrapper for mwclient with improvements for 3rd party wikis"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"},
  {name = "Tim Holzheim", email = "tim.holzheim@rwth-aachen.de"}
]
maintainers = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" },
  {name = "Tim Holzheim", email = "tim.holzheim@rwth-aachen.de"}
]
readme = "README.md"
license= "Apache-2.0"
dependencies = [
	# https://pypi.org/project/pybasemkit/
	# https://github.com/WolfgangFahl/pybasemkit
	# Python base module kit: YAML/JSON I/O, structured logging, CLI tooling, shell execution, and remote pydevd debug support.
	"pybasemkit>=0.1.3",
	# https://pypi.org/project/pyLodStorage/
    "pyLodStorage>=0.17.0",
	# https://pypi.org/project/GitPython/
	'gitpython',
	# https://pypi.org/project/Jinja2/
    'jinja2',
    # https://pypi.org/project/pywikibot/
    'pywikibot>=8.2',
    #https://pypi.org/project/requests/
    'requests>=2.32.5',
    # https://pypi.org/project/pycryptodome/
    'pycryptodome>=3.20.0',
    # https://pypi.org/project/mwclient/
    'mwclient>=0.11.0',
    # https://pypi.org/project/mwparserfromhell/
    'mwparserfromhell>=0.7.2',
    # https://pypi.org/project/wikitextparser/
    'wikitextparser>=0.56.4',
    # mysql-connector-python 8.3.0
    'mysql-connector-python>=9.4.0',
	# https://pypi.org/project/bcrypt/
	"bcrypt>=4.3.0",
	# https://pypi.org/project/shutup/
	"shutup>=0.2.0",
	# https://pypi.org/project/tqdm/
	"tqdm>=4.67.1"
]

requires-python = ">=3.10"
classifiers=[
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License"
    ]
dynamic = ["version"]
[tool.hatch.version]
path = "wikibot3rd/__init__.py"

[project.urls]
homepage = "https://wiki.bitplan.com/index.php/Py-3rdparty-mediawiki"
documentation = "https://wiki.bitplan.com/index.php/Py-3rdparty-mediawiki"
repository = "https://wiki.bitplan.com/index.php/Py-3rdparty-mediawiki"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
only-include = ["wikibot3rd","wikibot3rd_examples"]

[tool.hatch.build.targets.wheel.sources]
"wikibot3rd" = "wikibot3rd"
"wikibot3rd_examples" = "wikibot3rd_examples"


[project.scripts]
wikibackup = "wikibot3rd.wikipush:mainBackup"
wikiedit = "wikibot3rd.wikipush:mainEdit"
wikinuke = "wikibot3rd.wikipush:mainNuke"
wikipush = "wikibot3rd.wikipush:mainPush"
wikiquery = "wikibot3rd.wikipush:mainQuery"
wikiupload = "wikibot3rd.wikipush:mainUpload"
wikirestore = "wikibot3rd.wikipush:mainRestore"
wikiuser = "wikibot3rd.wikiuser_cmd:main"
